Move old stuff to one dir
